* Docker image has prod build * Docker image prints version on startup * Build command now creates a self-contained archive * Builds nightly (version with git  hash) or prod * Fixes prisma generate being run too late

+++ scripts/build-image.sh
@@ -1,12 +1,22 @@
 #!/bin/bash
 
-SPLIIT_APP_NAME=$(node -p -e "require('./package.json').name")
-SPLIIT_VERSION=$(node -p -e "require('./package.json').version")
+source scripts/set-env.sh
 
-# we need to set dummy data for POSTGRES env vars in order for build not to fail
-docker buildx build \
-    -t ${SPLIIT_APP_NAME}:${SPLIIT_VERSION} \
-    -t ${SPLIIT_APP_NAME}:latest \
-    .
+if [ ! -d "${SPLIIT_DEST_DIR}" ]; then
+    echo "Build folder ${SPLIIT_DEST_DIR} not found"
+    exit 1
+fi
+
+mkdir -p ${SPLIIT_DIST_FOLDER}
+
+docker buildx   build \
+                --no-cache \
+                --build-arg "SPLIIT_APP_NAME=${SPLIIT_APP_NAME}" \
+                --build-arg "SPLIIT_VERSION=${SPLIIT_VERSION}" \
+                --build-arg "SPLIIT_DEST_DIR=${SPLIIT_DEST_DIR}" \
+                --build-arg "SPLIIT_NODE_VERSION=${SPLIIT_NODE_VERSION}" \
+                -t ${SPLIIT_APP_NAME}:${SPLIIT_VERSION} \
+                -t ${SPLIIT_APP_NAME}:latest .
+docker save ${SPLIIT_APP_NAME}:${SPLIIT_VERSION} > ./${SPLIIT_DIST_FOLDER}/${SPLIIT_APP_NAME}-${SPLIIT_VERSION}-docker.tar.gz
 
 docker image prune -f
